+++ Realm/RLMUser.mm
@@ -85,6 +85,10 @@ using namespace realm;
     _user->log_out();
 }
 
+- (BOOL)isLoggedIn {
+    return _user->is_logged_in();
+}
+
 - (void)invalidate {
     if (!_user) {
         return;
@@ -96,7 +100,7 @@ using namespace realm;
     std::stringstream s;
     s << RLMConvertRLMBSONToBson(partitionValue);
     NSString *encodedPartitionValue = [@(s.str().c_str()) stringByAddingPercentEncodingWithAllowedCharacters:[NSCharacterSet URLPathAllowedCharacterSet]];
-    return [[NSString alloc] initWithFormat:@"%@/%@", [self identity], encodedPartitionValue];
+    return [[NSString alloc] initWithFormat:@"%@/%@", [self identifier], encodedPartitionValue];
 }
 
 - (nullable RLMSyncSession *)sessionForPartitionValue:(id<RLMBSON>)partitionValue {
@@ -123,7 +127,7 @@ using namespace realm;
     return [buffer copy];
 }
 
-- (NSString *)identity {
+- (NSString *)identifier {
     if (!_user) {
         return nil;
     }
@@ -138,7 +142,7 @@ using namespace realm;
     auto identities = _user->identities();
     for (auto& identity : identities) {
         [buffer addObject: [[RLMUserIdentity alloc] initUserIdentityWithProviderType:@(identity.provider_type.c_str())
-                                                                            identity:@(identity.id.c_str())]];
+                                                                          identifier:@(identity.id.c_str())]];
     }
 
     return [buffer copy];
@@ -192,7 +196,7 @@ using namespace realm;
     });
 }
 
-- (RLMAPIKeyAuth *)apiKeyAuth {
+- (RLMAPIKeyAuth *)apiKeysAuth {
     return [[RLMAPIKeyAuth alloc] initWithApp: _app];
 }
 
@@ -270,10 +274,10 @@ using namespace realm;
 @implementation RLMUserIdentity
 
 - (instancetype)initUserIdentityWithProviderType:(NSString *)providerType
-                                        identity:(NSString *)identity {
+                                      identifier:(NSString *)identifier {
     if (self = [super init]) {
         _providerType = providerType;
-        _identity = identity;
+        _identifier = identifier;
     }
     return self;
 }